requirements.txt
asyncpg==0.29.0
websockets==12.0
python-dotenv==1.0.1
orjson==3.10.7
pytz==2024.1
aiohttp==3.9.5
